Revert "Merge branch 'topbarFrontend' into sprint4"
This reverts commit f58bfde66721ba504ffef8cd82e4aca442abfc44, reversing
changes made to a80d17a49c0a5602fc4a2d41a115a9e698804b06.
