+++ WebServer/frontend/package.json
@@ -23,7 +23,6 @@
     "@angular/router": "~9.0.2",
     "@auth0/angular-jwt": "^4.0.0",
     "hls.js": "^0.13.2",
-    "material-design-icons": "^3.0.1",
     "ngx-videogular": "^9.0.1",
     "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
