+++ WebServer/frontend/src/app/app-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomepageComponent } from './homepage/homepage.component'
+import { LoginComponent } from './login/login.component'
 import { AboutComponent } from './about/about.component';
 
 const routes: Routes = [
@@ -8,6 +9,7 @@ const routes: Routes = [
     path: "",
     children: [
       { path: "", component: HomepageComponent },
+      { path: "login", component: LoginComponent },
       { path: "about", component: AboutComponent },
     ]
   },
