+++ WebServer/frontend/src/app/app.component.html
@@ -1,3 +1,2 @@
 <!-- This is the default page, if topbar or bottombar is needed for all pages, add it here -->
-<app-topbar></app-topbar>
 <router-outlet></router-outlet>
