+++ WebServer/frontend/src/app/app.module.ts
@@ -4,6 +4,7 @@ import { FormsModule } from "@angular/forms";
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './app-routing.module';
 import { HomepageComponent } from './homepage/homepage.component';
+import { LoginComponent } from './login/login.component';
 import { VideoplayerComponent } from './videoplayer/videoplayer.component';
 import { VgCoreModule, VgControlsModule, VgStreamingModule, VgOverlayPlayModule, VgBufferingModule } from 'ngx-videogular';
 import { ProfileService } from './_services/profile.service';
@@ -18,17 +19,16 @@ import { MatDialogModule } from '@angular/material/dialog';
 import { LoginService } from './_services/login.service';
 import { JwtHelperService, JWT_OPTIONS } from '@auth0/angular-jwt';
 import { AuthService } from './_services/auth.service';
-import { TopbarComponent } from './topbar/topbar.component';
 import { AboutComponent } from './about/about.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     HomepageComponent,
+    LoginComponent,
     VideoplayerComponent,
     SidemenuComponent,
     LoginDialog,
-    TopbarComponent,
     AboutComponent,
   ],
   imports: [
