+++ WebServer/frontend/src/app/sidemenu/sidemenu.component.html
@@ -1,5 +1,12 @@
-<h3 style="text-align: center;">Camera</h3>
+<h1 style="text-align: center;">ICU</h1>
 <ul style="padding-left: 0%;">
     <li *ngFor="let stream of streams" (click)="onClickStream(stream)"
         [ngClass]="{'selected': currentStream === stream}">{{ stream.label }}</li>
 </ul>
+<div style="display:flex; justify-content: center;">
+    <button mat-raised-button *ngIf="!loggedIn" (click)="openDialog()">Login</button>
+    <div *ngIf="loggedIn">
+        <p>{{this.email}}</p>
+        <button>Add Cam</button>
+    </div>
+</div>
