--- WebServer/frontend/src/app/topbar/topbar.component.html
@@ -1,15 +1,0 @@
-<mat-toolbar style="height: 3rem;" >
-    <h1>ICU</h1>
-    <span class="fill-space"></span>
-    <button mat-raised-button style="margin-right: 1rem;" (click)="doToggleAbout()">{{isHomepage? 'About': 'Stream'}}</button>
-    <button mat-raised-button *ngIf="!isLoggedIn" (click)="openDialog()">Login</button>
-    <button mat-button [matMenuTriggerFor]="menu" *ngIf="isLoggedIn">{{email}}
-        <i class="material-icons" style="vertical-align: middle;">arrow_drop_down</i>
-    </button>
-    <mat-menu #menu="matMenu">
-      <button mat-menu-item>Profile</button>
-      <button mat-menu-item *ngIf="isAdmin">Create User</button>
-      <button mat-menu-item (click)="logout()">Logout</button>
-    </mat-menu>
-    
-</mat-toolbar>
