--- WebServer/frontend/src/app/topbar/topbar.component.scss
@@ -1,5 +1,0 @@
-.fill-space{
-    /* This fills the remaining space, by using flexbox. 
-       Every toolbar row uses a flexbox row layout. */
-    flex: 1 1 auto;
-  }
