+++ WebServer/frontend/src/app/videoplayer/videoplayer.component.html
@@ -30,6 +30,7 @@
                [vgDash]="currentStream.source"
                [vgHls]="currentStream.source"
                id="singleVideo"
+               autoplay
                crossorigin>
         </video>
