+++ WebServer/frontend/src/app/videoplayer/videoplayer.component.scss
@@ -6,30 +6,30 @@ vg-player, video {
 //Add this if we really really want 16:9 aspect ratio.
 //(if not here it just scale with the video can give other problems tho)
 
-.videogular-container {
-	width: 100%;
-	height: 320px;
-	margin: auto;
-	overflow: hidden;
-}
+// .videogular-container {
+// 	width: 100%;
+// 	height: 320px;
+// 	margin: auto;
+// 	overflow: hidden;
+// }
  
-@media (min-width: 1200px) {
-	.videogular-container {
-		width: 1170px;
-		height: 658.125px;
-	}
-}
+// @media (min-width: 1200px) {
+// 	.videogular-container {
+// 		width: 1170px;
+// 		height: 658.125px;
+// 	}
+// }
  
-@media (min-width: 992px) and (max-width: 1199px) {
-	.videogular-container {
-		width: 940px;
-		height: 528.75px;
-	}
-}
+// @media (min-width: 992px) and (max-width: 1199px) {
+// 	.videogular-container {
+// 		width: 940px;
+// 		height: 528.75px;
+// 	}
+// }
  
-@media (min-width: 768px) and (max-width: 991px) {
-	.videogular-container {
-		width: 728px;
-		height: 409.5px;
-	}
-}
+// @media (min-width: 768px) and (max-width: 991px) {
+// 	.videogular-container {
+// 		width: 728px;
+// 		height: 409.5px;
+// 	}
+// }
