+++ WebServer/frontend/src/styles.scss
@@ -1,6 +1,5 @@
 /* You can add global styles to this file, and also import other style files */
 @import '~@angular/material/prebuilt-themes/purple-green.css';
-@import '~material-design-icons/iconfont/material-icons.css';
 
 html, body{
     background-color: gray;
